# src/main/resources/swagger.api.def.json
{
  "swagger": "2.0",
  "info": {
    "description": "REST CRUD Interface",
    "title": "REST CRUD Interface",
    "version": "v1"
  },
  "host": "crud.org",
  "schemes": [
    "http"
  ],
  "basePath": "/",
  "paths": {
    "/api/event": {
      "get": {
        "summary": "List Events",
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "type":"array",
              "items": {
                "$ref": "#/definitions/Event"
              }
            }
          }
        }
      },
      "post":{
        "summary": "Creates an Event",
        "consumes":[ "application/x-www-form-urlencoded" ],
        "produces":[ "application/json" ],
        "operationId":"newEvent",
        "parameters":[{
          "in":"body",
          "name":"newEvent",
          "required":true,
          "schema": {
            "$ref": "#/definitions/Event"
          }
        }],
        "responses": {
          "200":{
            "description":"created",
            "schema": {
              "type": "integer",
              "format":"in64"
            }
          },
          "400":{
            "description":"invalid event"
          }
        }
      }
    },
    "/api/event/{id}": {
      "get": {
        "summary": "Get Events By Id",
        "parameters": [{
          "name":"id",
          "in":"path",
          "required":true,
          "type":"integer",
          "format":"in64"
        }],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/Event"
            }
          }
        }
      },
      "put" : {
        "summary": "Updates an Event",
        "consumes":[ "application/x-www-form-urlencoded" ],
        "produces":[ "application/json" ],
        "operationId":"editEvent",
        "parameters":[{
          "name":"id",
          "in":"path",
          "required":true,
          "type":"integer",
          "format":"in64"
          },
          {
          "in":"body",
          "name":"event",
          "required":true,
          "schema": {
            "$ref": "#/definitions/Event"
          }
        }],
        "responses": {
          "200":{
            "description":"updated",
            "schema": {
              "type": "boolean",
              "default":false
            }
          },
          "400":{
            "description":"invalid event"
          }
        }
      },
      "delete" : {
        "summary": "Deletes an Event",
        "produces":[ "application/json" ],
        "operationId":"delEvent",
        "parameters":[{
          "name":"id",
          "in":"path",
          "required":true,
          "type":"integer",
          "format":"in64"
        }],
        "responses": {
          "200":{
            "description":"deleted",
            "schema": {
              "type": "boolean",
              "default":false
            }
          },
          "400":{
            "description":"invalid event id"
          }
        }
      }
    }
  },
  "definitions":
  {
    "Event":
    {
      "type":"object",
      "properties":
      {
        "id" :
        {
          "type":"integer",
          "format":"int64"
        },
        "Name" :
        {
          "type":"string"
        },
        "Desc" :
        {
          "type":"string"
        },
        "Place" :
        {
          "type":"string"
        },
        "Date" :
        {
          "type":"string",
          "format":"date-time"
        }
      }
    }
  }
}